Use `currentTarget` instead to make sure listener fires on element
When the sharing handler attached to an element containing another
element (e.g. button containing an icon), the sharing functionality
wouldn't fire because the handler was firing on the child elements
instead of the sharer element.

For example, this wouldn't work:

```
<button class="sharer button" data-sharer="twitter" data-title="Checkout
Sharer.js!" data-via="ellisonleao" data-hashtags="awesome, sharer.js"
data-url="https://ellisonleao.github.io/sharer.js/"><span>Share on
Twitter</span></button>
```

This commit resolves that issue by using `Event.currentTarget` instead
of `Event.target`. From the [docs]:

> Identifies the current target for the event, as the event traverses
> the DOM. It always refers to the element the event handler has been
> attached to as opposed to event.target which identifies the element on
> which the event occurred.

This always gives us the element that the sharer was attached to as
opposed to any children elements, which is what we want.

[docs]:
https://developer.mozilla.org/en-US/docs/Web/API/Event/currentTarget